#6. Moving existing components.
