--- src/components/index.tsx
@@ -1,18 +1,0 @@
-export { LmPane } from "./LmPane";
-// export { Pane2 } from "./Pane2";
-export { FcPane } from "./FcPane";
-export { PreviewPane } from "./PreviewPane";
-export { LmDropdown } from "./LmDropdown";
-export { FcDropdown } from "./FcDropdown";
-// export { ConceptLabel } from "./ConceptLabel";
-export { CardDisplay } from "./CardDisplay";
-export { CardAdd } from "./CardAdd";
-export { CardEdit } from "./CardEdit";
-export { McqAnswerField } from "./McqAnswerField";
-export { QaAnswerField } from "./QaAnswerField";
-// export { DeleteButton } from "./DeleteButton";
-// export { SaveButton } from "./SaveButton";
-export { LoginButton } from "./LoginButton";
-export { LogoutButton } from "./LogoutButton";
-export { Profile } from "./Profile";
-export { LandingPage } from "./LandingPage";
